Move shorthand into NetCore/Short.hs
This move is necessary in preparation for building a better pretty-printer
because the pretty-printing modules use some of the same shorthand infix
combinators as we do.  This also helped find places where some code was
depending on NetCore/API where it should have been depending on NetCore as a
whole, which is preferable because it lets us do this kind of refactor without
breaking things.

Now, API very nearly only contains datatype and their instances, and Short
contains functions and combinators.