Demonstrate that eliminating Patterns should be performant.
I'm going to remove the PoPattern variant from Predicates, and simply
have equality constraints between header names and values (and prefixes).

This commit changes the reducer to "explode" a pattern with N exact matches
into an intersection of N patterns with 1 exact match each.

This commit also has a new reducer: it first transforms to disjunctive-normal
form, then simplifies each conjunct.

Here are performance numbers with the old technique:

  arjun@linux-vm:~/src/netcore$ time cabal test --test-option=sat
  Running 1 test suites...
  Test suite frenetic-tests: RUNNING...
  Test suite frenetic-tests: PASS
  Test suite logged to: dist/test/netcore-1.0.0-frenetic-tests.log
  1 of 1 test suites (1 of 1 test cases) passed.

  real	1m5.237s
  user	0m30.438s
  sys	0m1.044s

The new technique is about 13% slower:

  arjun@linux-vm:~/src/netcore$ time cabal test --test-option=sat
  Running 1 test suites...
  Test suite frenetic-tests: RUNNING...
  Test suite frenetic-tests: PASS
  Test suite logged to: dist/test/netcore-1.0.0-frenetic-tests.log
  1 of 1 test suites (1 of 1 test cases) passed.

  real	1m10.641s
  user	0m34.934s
  sys	0m0.940s

But, I suspect it will be faster when I actually remove patterns, instead
of simply exploding them as in this commit.