Example program demonstrates composition of packet-output channels.
This is a sloppy, late-night demo for SendPacket. In PacketOut.hs, the thread
writing to pktOutChan produces a bogus ARP packet on both ports 1 and 2.
However, we use policy restriction to restrict the output to port 1.

To run it, start the PktOut controller:

./dist/build/frenetic-example-suite/frenetic-example-suite --verbosity=DEBUG --pktout

Start Mininet like so:

sudo mn --switch=user --controller=remote --mac --arp

Either trust the debugging output or use Wireshark.