Fixed bug that was causing the learning switch to always flood and never learn.
The bug was caused by (isQuery GetPkts = False). This in turn caused toCtlr to return [] in actnTranslate!

In addition, actnTranslate looked a little bogus as acts was defined but never used. (I suspect this means unimplementable modifications were not being handled correctly, but I didn't observe this in any of my tests.)